+++ sample-apps/react-native/expo-video-sample/ios/expovideosample/AppDelegate.mm
@@ -1,7 +1,9 @@
 #import "AppDelegate.h"
+#import "StreamVideoReactNative.h"
 #import "RNVoipPushNotificationManager.h"
 #import <PushKit/PushKit.h>
 #import "RNCallKeep.h"
+#import <Firebase/Firebase.h>
 
 #import <React/RCTBundleURLProvider.h>
 #import <React/RCTLinkingManager.h>
@@ -18,6 +20,9 @@
     @"supportsVideo": @YES,
     @"includesCallsInRecents": @NO,
   }];
+// @generated begin @react-native-firebase/app-didFinishLaunchingWithOptions - expo prebuild (DO NOT MODIFY) sync-ecd111c37e49fdd1ed6354203cd6b1e2a38cccda
+[FIRApp configure];
+// @generated end @react-native-firebase/app-didFinishLaunchingWithOptions
   self.moduleName = @"main";
 
   // You can add your custom initial props in the dictionary below.
@@ -82,6 +87,11 @@
   NSDictionary *stream = payload.dictionaryPayload[@"stream"];
   NSString *uuid = [[NSUUID UUID] UUIDString];
   NSString *createdCallerName = stream[@"created_by_display_name"];
+  NSString *cid = stream[@"call_cid"];
+
+  [StreamVideoReactNative registerIncomingCall:cid uuid:uuid];
+
+  [RNVoipPushNotificationManager addCompletionHandler:uuid completionHandler:completion];
 
   // display the incoming call notification
   [RNCallKeep reportNewIncomingCall: uuid
@@ -95,7 +105,7 @@
                  supportsUngrouping: YES
                         fromPushKit: YES
                             payload: stream
-              withCompletionHandler: completion];
+              withCompletionHandler: nil];
 }
 
 @end