+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.4.18](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.17...@stream-io/video-react-bindings-0.4.18) (2024-04-23)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.7.7`
+### [0.4.17](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.16...@stream-io/video-react-bindings-0.4.17) (2024-04-23)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.7.6`
+### [0.4.16](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.15...@stream-io/video-react-bindings-0.4.16) (2024-04-23)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.7.5`
 ### [0.4.15](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.4.14...@stream-io/video-react-bindings-0.4.15) (2024-04-17)
 
 ### Dependency Updates